+++ src/design/plone/contenttypes/upgrades.py
@@ -272,6 +272,7 @@ def to_1013(context):
 
 
 def to_1014(context):
+    update_types(context)
     portal_types = api.portal.get_tool(name="portal_types")
     portal_types["Bando"].behaviors = tuple(
         [
@@ -279,4 +280,18 @@ def to_1014(context):
             for x in portal_types["Bando"].behaviors
             if x != "design.plone.contenttypes.behavior.argomenti"
         ]
+    )
+
+
+def to_1015(context):
+    update_types(context)
+
+    # cleanup trasparenza behavior from CTs
+    portal_types = api.portal.get_tool(name="portal_types")
+    service_behaviors = portal_types["Servizio"].behaviors
+    to_remove = [
+        "design.plone.contenttypes.behavior.trasparenza",
+    ]
+    portal_types["Servizio"].behaviors = tuple(
+        [x for x in service_behaviors if x not in to_remove]
     )